+++ jaws-core/src/test/java/org/hongxi/jaws/cluster/loadbalance/LeastActiveLoadBalanceTest.java
@@ -119,8 +119,8 @@ class LeastActiveLoadBalanceTest {
             @Override public long getRequestId() { return 0; }
             @Override public int getRetries() { return 0; }
             @Override public void setRetries(int retries) {}
-            @Override public int getSerializationNumber() { return 0; }
-            @Override public void setSerializationNumber(int number) {}
+            @Override public byte getSerializationNumber() { return 0; }
+            @Override public void setSerializationNumber(byte number) {}
         };
     }
 }
