+++ jaws-core/src/test/java/org/hongxi/jaws/cluster/loadbalance/RoundRobinLoadBalanceTest.java
@@ -110,8 +110,8 @@ class RoundRobinLoadBalanceTest {
             @Override public long getRequestId() { return 0; }
             @Override public int getRetries() { return 0; }
             @Override public void setRetries(int retries) {}
-            @Override public int getSerializationNumber() { return 0; }
-            @Override public void setSerializationNumber(int number) {}
+            @Override public byte getSerializationNumber() { return 0; }
+            @Override public void setSerializationNumber(byte number) {}
         };
     }
 }
